requirements-test.txt
# Testing dependencies
pytest>=7.0.0
pytest-mock>=3.6.1
pytest-cov>=4.0.0
pytest-html>=3.1.0
moto>=4.0.0  # For AWS service mocking
boto3>=1.26.0
botocore>=1.29.0
PyJWT==2.8.0
ulid-py==1.1.0
requests>=2.31.0  # For E2E integration tests
